+++ app/src/marketplace/components/ActionBar/FilterSelector/index.jsx
@@ -6,9 +6,9 @@ import cx from 'classnames'
 import MediaQuery from 'react-responsive'
 import breakpoints from '$app/scripts/breakpoints'
 
-import type { AnyFilter } from '../../../flowtype/product-types'
+import type { AnyFilter } from '$mp/flowtype/product-types'
+import dropdownStyles from '$shared/components/DropdownActions/dropdownActions.pcss'
 
-import dropdownStyles from '$shared/components/Dropdown/dropdown.pcss'
 import styles from './filterSelector.pcss'
 import FilterModal from './FilterModal'
 import FilterDropdown from './FilterDropdown'
