+++ app/src/shared/components/index.js
@@ -2,7 +2,7 @@
 
 export { default as Page } from './Page'
 export { default as Button } from './Button'
-export { default as Dropdown } from './Dropdown'
+export { default as DropdownActions } from './DropdownActions'
 export { default as Footer } from './Footer'
 export { default as FooterColumn } from './Footer/FooterColumn'
